`fix` change data type

+++ features/content/play/src/main/java/com/tokopedia/play/util/logger/PlayLog.kt
@@ -13,7 +13,7 @@ interface PlayLog {
     )
 
     fun logDownloadSpeed(
-        downloadSpeed: Int
+        downloadSpeed: Float
     )
 
     fun logBufferEvent(
